fix stream access token expired
